# Performance improvements for add_method_to_attributes!