@rafaelfranca 2017-11-07
Looks good to me. Can you squash your commits in only one?