@eugeneius 2017-11-07
Did you investigate @matthewd's suggestion here? https://github.com/rails/rails/pull/30963#discussion_r146381317

It seems like a simpler way to save the same number of allocations.