@oniofchaos 2017-11-09
Took another look at that comment and tested it out and looks like it provides a 2x speedup compared to my current solution for non-get arguments such as `delete` but a 10x slowdown for `get`. Not sure if that's worth? Or maybe there's a further optimization to make based on using `match`?

Updated benchmarking code and results:
```ruby
# frozen_string_literal: true
begin
  require "bundler/inline"
rescue LoadError => e
  $stderr.puts "Bundler version 1.10 or later is required. Please update
                your Bundler"
  raise e
end

gemfile(true) do
  source "https://rubygems.org"

  gem "benchmark-ips"
  gem "rails"
end

def allocate_count
  GC.disable
  before = ObjectSpace.count_objects
  yield
  after = ObjectSpace.count_objects
  after.each { |k,v| after[k] = v - before[k] }
  after[:T_HASH] -= 1 # probe effect - we created the before hash.
  GC.enable
  result = after.reject { |k,v| v == 0 }
  GC.start
  result
end

@html_options = {}

def master_version
  if @method && @method.to_s.downcase != "get" && @html_options["rel".freeze] !~ /nofollow/
    @html_options["rel".freeze] = "#{@html_options["rel".freeze]} nofollow".lstrip
  end
  @html_options["data-method".freeze] = @method
end

def fast_version
  if method_not_get_method?(@method) && @html_options["rel".freeze] !~ /nofollow/
    @html_options["rel".freeze] = "#{@html_options["rel".freeze]} nofollow".lstrip
  end
  @html_options["data-method".freeze] = @method
end

def match_version
  if @method && /\Aget\z/i.match?(@method) && @html_options["rel".freeze] !~ /nofollow/
    @html_options["rel".freeze] = "#{@html_options["rel".freeze]} nofollow".lstrip
  end
  @html_options["data-method".freeze] = @method
end

def match_version_with_const
  comp_method = (STRINGIFIED_COMMON_METHODS[@method] || @method)
  if @method && /\Aget\z/i.match?(comp_method) && @html_options["rel".freeze] !~ /nofollow/
    @html_options["rel".freeze] = "#{@html_options["rel".freeze]} nofollow".lstrip
  end
  @html_options["data-method".freeze] = @method
end

STRINGIFIED_COMMON_METHODS = {
  get:    'get',
  delete: 'delete',
  patch:  'patch',
  post:   'post',
  put:    'put',
}.freeze

def method_not_get_method?(method)
  return false unless method
  (STRINGIFIED_COMMON_METHODS[method] || method.to_s.downcase) != 'get'
end

def test
  puts "master_version"
  puts allocate_count { 1000.times { master_version } }
  puts "fast_version"
  puts allocate_count { 1000.times { fast_version } }
  puts "match_version"
  puts allocate_count { 1000.times { match_version } }
  puts "match_version_with_const"
  puts allocate_count { 1000.times { match_version_with_const } }

  Benchmark.ips do |x|
    x.report("master_version") { master_version }
    x.report("fast_version")     { fast_version }
    x.report("match_version")     { match_version }
    x.report("match_version_with_const")     { match_version_with_const }
    x.compare!
  end
end

puts 'get'
@method = :get

test

puts 'delete'
@method = :delete

test
```ruby
get
master_version
{:FREE=>-1932, :T_STRING=>2052}
fast_version
{:FREE=>-1}
match_version
{:FREE=>-5003, :T_STRING=>1003, :T_MATCH=>1999, :T_IMEMO=>2000}
match_version_with_const
{:FREE=>-5003, :T_STRING=>1002, :T_MATCH=>2000, :T_IMEMO=>2000}
Warming up --------------------------------------
      master_version   117.834k i/100ms
        fast_version   170.545k i/100ms
       match_version    33.021k i/100ms
match_version_with_const
                        35.210k i/100ms
Calculating -------------------------------------
      master_version      2.513M (± 9.4%) i/s -     12.490M in   5.017587s
        fast_version      4.038M (± 8.2%) i/s -     20.124M in   5.022930s
       match_version    385.497k (± 9.5%) i/s -      1.915M in   5.011259s
match_version_with_const
                        398.243k (±11.8%) i/s -      1.972M in   5.024023s

Comparison:
        fast_version:  4037895.6 i/s
      master_version:  2512706.7 i/s - 1.61x  slower
match_version_with_const:   398243.2 i/s - 10.14x  slower
       match_version:   385496.7 i/s - 10.47x  slower

delete
master_version
{:FREE=>-5002, :T_STRING=>3001, :T_MATCH=>1000, :T_IMEMO=>1000}
fast_version
{:FREE=>-3002, :T_STRING=>1001, :T_MATCH=>1000, :T_IMEMO=>1000}
match_version
{:FREE=>-1001, :T_IMEMO=>1000}
match_version_with_const
{:FREE=>-1001, :T_IMEMO=>1000}
Warming up --------------------------------------
      master_version    43.377k i/100ms
        fast_version    58.810k i/100ms
       match_version    98.709k i/100ms
match_version_with_const
                        96.942k i/100ms
Calculating -------------------------------------
      master_version    629.208k (± 9.9%) i/s -      3.123M in   5.009754s
        fast_version    715.365k (± 9.1%) i/s -      3.587M in   5.052967s
       match_version      1.466M (± 6.0%) i/s -      7.304M in   5.000822s
match_version_with_const
                          1.437M (± 7.5%) i/s -      7.174M in   5.025284s

Comparison:
       match_version:  1466121.0 i/s
match_version_with_const:  1436722.2 i/s - same-ish: difference falls within error
        fast_version:   715365.0 i/s - 2.05x  slower
      master_version:   629208.3 i/s - 2.33x  slower
```